- Merged libmjollnir with modflow ! : 	* modflow doesnt exist anymore 	* Big cleanup of modflow API which is now in mjollnir 	* Moved btree functions from libelfsh to mjollnir 	* Prefixed in a unified way the mjollnir/modflow functions and data structures 	* mjollnir now compiles by default without having to enable it
The block format of mjollnir is still not merged with the one of modflow
but its work in progress and will be commited soon. Both analysis still works
using "analyse" and "flow" commands. The flowjack, graph, flow, analyse,
rename, unstrip, etc commands function (cmd_*) are now all in vm/

- Cleaned a lot the code of disasm.c, now we have many more small functions. Look
at it, its way better now.

TODO for the mjollnir merge :

- Cleanup the funcptr hijack function
- Make .control format compatible with mjollnir blocks
- Merge block format
- Merge "analyse" and "flow" commands

-may


git-svn-id: http://svn.eresi-project.org/svn/trunk@255 cfe96d8e-f5de-dc11-b1ed-0017314a9460