Libaspect is now merged. Files were moved and a new library is now available : libaspect/libaspect.so. -I and -L -l added in the code, the testsuite is not yet up to date except the relinject/etrel_extplt/ (see next point)
Librevm : created the directory and added in the configure, but not
yet enabled as this modularisation is a little bit more subtle to do, so
we dont have files in librevm directory yet.

Strauss / Heroine : you can now use vectors and hashes in libasm ! Just
add the correct -L -l and -I in your Makefiles !

Next commits to follow.

-may


Cvs : ----------------------------------------------------------------------


git-svn-id: http://svn.eresi-project.org/svn/trunk@286 cfe96d8e-f5de-dc11-b1ed-0017314a9460