Correct hash_empty() fonction that use elemnbr and not size element which can cause a Floating point fault in hash_get_head():
  return (&h->ent[index % h->size]);

if hash_empty() occur when the table is empty, elemnbr is 0 then h->size will
be at 0, index % 0 create the error.

I didn't add a check on hash_init because noone check hash_init return, so it
would be useless, but it should be done as calloc can return a pointer if size is 0,
just to ensure that it can be free.


git-svn-id: http://svn.eresi-project.org/svn/trunk@422 cfe96d8e-f5de-dc11-b1ed-0017314a9460