- Added one point in the coding style documentation
- Added vectors and hash tables as real scripting objects (to use with
set/get/add/del, etc). NOTE: only the API was done, it is not yet in
the language)

- Added multidimensional arrays to the type system.

- Corrected behavior when a structure type was refused when it had only
one field : now this is allowed and can be used for doing type aliases
for instance.

- Extended a little bit the testsuite script of the type system

More commits to come for potential bugs, and the final integration of
vectors/hash in the language.

-may


git-svn-id: http://svn.eresi-project.org/svn/trunk@433 cfe96d8e-f5de-dc11-b1ed-0017314a9460