Corrected some bugs and "finished" porting mjollnir to SPARC
'analyse' runs well until it hits mjr_function_flow_parents_save(), where it tries to copy some data from the _real_ address of the function (ie. the address where it would be if the binary was loaded)
that's a bug? maybe the function intended to copy data from the buffer storing the file contents, or just to copy the address itself?


git-svn-id: http://svn.eresi-project.org/svn/trunk@439 cfe96d8e-f5de-dc11-b1ed-0017314a9460