We switch to 0.77. This is major update. Only e2dbg, librevm, and libui are affected.
- Appeared a new library : libe2dbg. This holds the (embedded) server
part of the debugger.

- Moved files from libui to librevm for beeing able not to link with
libui if we dont want readline (ex: in the embedded debugger)

- Created more subdirectories for librevm. Now have api/ and io/
in more than cmd/ elf/ vm/ and lang/

- Create a separate header file for librevm-io (see librevm/include/)

- Did a huge code cleaning in libui and the network code, prefixed
all readline related functions by readln_, that stand inside libui/

- Changed the code in e2dbg/ for a client-like code (like etrace and
elfsh). MXATONE: I am doing some modification in librevm so that we
can give the first parameter as a ELF binary (and not a script file).
Dont worry about that point now, you should better focuss on the
injection bug that happens on ssh since its the only thing remaining
for you until the pre-release !

This will serve both inside e2dbg and etrace, but elfsh will stay as
it, so that it can continue to be the script interpreter.

I commit it because it should not affect anyone's work since I work
on the debugger and the language alone, BUT:

- dont try to use the debugger now

- dont try to use the --enable-network option now

- some scripts are failing (but nothing big is broken, all regular
scripts are working, its just some precise points which are broken, but
not for long)

Please everyone update, that will avoid losing modifications. I ported
all mxatone's modification to the new code too, so you should not have
any bad surprise.


git-svn-id: http://svn.eresi-project.org/svn/trunk@492 cfe96d8e-f5de-dc11-b1ed-0017314a9460