Ok, so this is the first commit of the new container-based libmjollnir It may not function accurately, so it needs lots of improvements, but it doesn't give any obvious segfaults
One major TODO is the fact that we need to rewrite some parts of the code that used the 'type'
field of mjrblock_t, which no longer exists. We should be able to infere the same information
using the 'type' member of the mjrlink_t structure. Anyway, some of the code that uses the old
type field is commented and maked with "FIXME", some of it was simply removed from the code and
we need to compare with the old versions in the cvs to see if there's any code that shouldn't
be removed (i'm sorry but i really thought that the type information wasn't necessary anymore).

There's also a lot of things to be commented using doxygen, and many functions could be merged.
I'm sorry if the indentation looks bad, i tried to use int3's vimrc, but i also discovered
that setting the variable tab-width to 2 in emacs, makes things look much better

So, these are the biggest changes:
- We now use containers almost everywhere
- We have a container api (make sure you try to understand the 'id' stuff)
- The block on-disk format now has a section to store link information too, like function already had
- The sections now are name .edfmt.blocks, .edfmt.bcontrol, .edfmt.functions, .edfmt.fcontrol (*control are link info)
- altdata for these sections above hold nothing... the hash tables of functions and blocks do all the job
- blklist, from the context structure, is gone

Cheers,

    Strauss


git-svn-id: http://svn.eresi-project.org/svn/trunk@505 cfe96d8e-f5de-dc11-b1ed-0017314a9460