- Completed man page (See doc/elfsh.1)
Mxatone: traces, versions command (verdef/verneed/version), and hash command
are not in the man page. Can you complete it ? Once  this is in, everything
is in (actually just the hash table and vector objects syntax will have to be
filled, but all commands are there). You may also need to complete the object
format documentation too for versions/verneed/verdef/elfhash objects that
you added in the scripting language.

I personally use the tool "manedit" (it sucks but is better than nothing),
anyone knows a better manual page editor ?

-may


git-svn-id: http://svn.eresi-project.org/svn/trunk@629 cfe96d8e-f5de-dc11-b1ed-0017314a9460