Major update : we switch to version 0.79 (alpha 1)
MAIN:

- Factored arithmetic commands to create a
  new API : vm_revmobj_set()

- Added typed variables declaration with neat syntax !

As suggested thorkill, the type name must be the name
of the command which declare the type. My solution was
simple : do a unique cmd_declare() and each type we
declare a variable (starting by its type name on the line..),
this function is called. The type of the variable to be
declared is then the curcmd->name itself. The name of
the variable is curcmd->param[0]. This systems work very
well and give us a very neat syntax for variable declarations.

- Added a recursive procedure that is capable of initializing
typed variables declared as such using a nice syntax.

Example:

type struct0 = field0:long
type struct1 = field1:byte field2:long
type struct2 = field3:struct1 field4:long field5:struct1

struct0 mystruct0 = (field0:42)
struct1 mystruct1 = (field1:42,field2:43)
struct2 mystruct2 = (field3(field1:42, field2:43), field4:44, field5(field1:45, field2:46))

This works well and declare 3 variables and assign them default values.

*** YOU DONT HAVE TO INITIALIZE ALL FIELDS, THOSE LEFT UNINITIALIZED ARE SET TO 0 ****

*** The name of the Root expressions does not have to be specified, but you can
also choose to put the expression type name : ExprType(field1:(....), field2:...., ...)
if you want to enforce type checking on variable initialization *****

- The internal ERESI type revmvar_t was renamed to revmannot_t.
This helps to make clear what is a real variable and what is
simply an annotation (in our case: annotations of types, but
in the future it can be anything because this new annot
structure aims at being more generic than revmvar_t)

- A new structure was created : revmexpr_t. This structure represents
a typed ERESI expression which handle the right values (on the right
of "=") when declaring variables. Expressions are trees of revmobj_t,
which turns into a perfect base for printing and comparison of
expressions, which is the prelude to program transformation using
match/case commands.


MISC:

- Added aliases for '(un)inform' command : (un)annotate (keep same parameters)
- Fixed Evarista prompt
- Fixed Etrace version (now uniform with ERESI version number)
- Fixed default REVM prompt to something like 'revm-...'
- Added a forgotten cp for make install of evarista in configure file
- Corrected warnings a little bit everywhere in the project

TODO:

- Setting of RAW typed variables and array variables
- Printing of expressions
- Comparison of expressions
- Create the object "list" and its API to be made accessible
  from the ERESI language.
- Make match/case to work in a small ERESI script that is
  parsing the CFG.


git-svn-id: http://svn.eresi-project.org/svn/trunk@710 cfe96d8e-f5de-dc11-b1ed-0017314a9460