- Introduced a new API for list/hash linearity. That allow to avoid bugs such as :
foreach $elem of $list[name]
  foreach $elem2 of $list[name]
    del $elem2 $list[name]
    print $elem
  forend
forend

e.g. object aliasing between $elem and $elem2 leading to bug.

The idea of linearity is that its not possible to iterate on a list
if its already being iterated. Of course, list of immediate values
are not concerned (foreach $i of 1 until 42). Only containers (hash
and lists) are subject to that bug.

The linearity API is only used in foreach.

- Added a new hash table : exprtypes_hash which returns the root type of
an expression giving its name. Before we had only exprs_hash which
returned the expression giving its name.

- Added an API of expressions for doing operations either by name or
by passing the expression structure in parameter directly

- Fixed indentation when printing types with their id (since the commit
of pouik). Everything is well indented again now.

- Finished to code the "case" command, including when a -list- of
LIR operation is given in destination (and not only 1 <-> 1 transformation)

- Added support for command "default" which is the default action to perform
when not matching anything in a match.

- Switched to version 0.8a3

WIP:

* elfsh/e2dbg is still not operational since the change of mxatone to use
hash tables for symbols. Please fix !

* match/case is not yet operational. Im working on it. Do not try to use
for now. Hopefully will work for next version.


git-svn-id: http://svn.eresi-project.org/svn/trunk@734 cfe96d8e-f5de-dc11-b1ed-0017314a9460