Sorry for the previous msg, was a little mistake from me. This is the real CVS message.
- Modified the LIR so that the dataflow is more easily computed. Now
we have more LIR instruction but the dataflow is made really really
easy. The previous backend is put in sparc2lir-OLD.esh for future
reference. The new backend is sparc2lir.esh.

- Created the code that do the intra-bloc dataflow analysis on the LIR
for any architecture, provided its code has been translated to LIR. It
is put in : lir-dataflow.esh. It is COMPLETE. Amazingly small, isnt it? :P

- Put all the LIR definitions in a script : lir-definition.esh

-> 0.8a15

NOTE: I have currently stopped to implement the ERESI interpreter
for focusing on my report (I just have 10 days to finish it). As
such, I am writing the ERESI code for analyzing programs but none
of it is operational at the moment. Use this as an academic/bibliographic
references only. The main reason is that the ERESI language should
remain independant of the implementation, so my report is not about
the implementation but the specification and semantics of the ERESI
language.


git-svn-id: http://svn.eresi-project.org/svn/trunk@768 cfe96d8e-f5de-dc11-b1ed-0017314a9460