This is a huge commit. Mostly evarista, librevm and libaspect are affected:
* Improved the function calling/returning in eresi language. Now it really
does what it is supposed to do (save parameters and restore them correctly
at any depth)

* The WHOLE code has been switched to use REVM expressions instead of
REVM objects. Note that a simple expression corresponds to an object.

This change is very interesting because it allows us to do an 'inform'
with some type and print the informed variable, and it will print
automatically in the good type format ! (without debug info)

* Fixed all the testsuite to make everything work again with the new
expression code. That was quite a big work.

* All of this result in a huge code cleanup. elfsh, e2dbg and etrace
have been tested/fixed and are known to work with the new expression
code. Evarista is still _NOT_ operational, but we are getting nearer
and nearer. This is mostly a backup commit because I still have things
to do to make everything perfect.

NOTE:
-----

- It is currently not possible to print a RAW object and print a whole
array (only array elements can be printed/assigned).

- It is currently not possible to initialize a subexpression with a
variable (but its possible to initialize a new expression with the
value of another). Of course, its possible to set a subexpression
with an expression, or inversely. My remark only holds for _initialization_
of complex variable.

-> Newly declared arrays or raw objects are initialized to 0 for now

-> Newly declared complex expressions have to be initialized by constants
for now. (see testsuite/testscripts/declare.esh for a full testsuite on
what is possible to do currently with (sub)expressions)

MY TODO:
--------

- Implement subexpressions initializations from existing expressions

- Continue to cleanup the inform system with expressions to reduce the
number of used hash tables

- Split librevm and create a new library : libstderesi to group all
commands, and let the lirevm code to be independant of the binary format,
just like libaspect is.

- Make program transformation to work in Evarista

- Start implementing dataflow analysis as specified by the ERESI language
reference (ask your copy if you dont have it)

- Be able to access lists and hash tables using $listname and $hashname
(additionally to $hash[name:elmname] and $list[name:elmname] currently)

- Implement full array initialization and printing

- Implement RAW object initialization and printing

- Improve printing of expressions (minor alignment problems when printing
complex typed expr)


git-svn-id: http://svn.eresi-project.org/svn/trunk@823 cfe96d8e-f5de-dc11-b1ed-0017314a9460