Summary: (0.82a3) Implemented software watchpoint. Allow to trace variables.
The command "watch" takes as much parameters as you want. Each parameter
is an expression that is evaluated. The result of evaluation is saved 
as an address to be traced.

Variable tracing is only active while doing stepping or while doing
itracing (step or itrace commands). You can add up to 50 watchpoints
(limit extensible if macro is changed in libe2dbg.h). 

This is the first step towards having fenris implemented in e2dbg
(and by extension, kedbg).



git-svn-id: http://svn.eresi-project.org/svn/trunk@1161 cfe96d8e-f5de-dc11-b1ed-0017314a9460